style: fix cs and type in `SSH2::get_stream_timeout()`

+++ phpseclib/Net/SSH2.php
@@ -3134,7 +3134,7 @@ class SSH2
      */
     private function get_stream_timeout()
     {
-        $sec = ini_get('default_socket_timeout');
+        $sec = (int) ini_get('default_socket_timeout');
         $usec = 0;
         if ($this->curTimeout > 0) {
             $sec = (int) floor($this->curTimeout);
